+++ CHANGELOG.md
@@ -59,6 +59,7 @@ If you like this project and find it useful, please consider giving it a star on
 ## [1.3.11] - 2024-07-08
 
 ### Added
+
 - [device]: Added addRequiredClusterServers and addOptionalClusterServers methods.
 - [frontend]: Added separated settings for the two logging systems (Matterbridge and Matter.js).
 
@@ -94,6 +95,7 @@ If you like this project and find it useful, please consider giving it a star on
 ## [1.3.9] - 2024-07-02
 
 ### Fixed
+
 - [matterbridge]: Fixed nodeLabel in childbridge mode
 - [matterbridge]: Fixed MeasurementClusters
 
@@ -104,6 +106,7 @@ If you like this project and find it useful, please consider giving it a star on
 ## [1.3.8] - 2024-07-01
 
 ### Fixed
+
 - [matterbridge]: Fixed crash in childbridge mode
 
 <a href="https://www.buymeacoffee.com/luligugithub">
@@ -113,6 +116,7 @@ If you like this project and find it useful, please consider giving it a star on
 ## [1.3.7] - 2024-06-30
 
 ### Added
+
 - [matter.js]: Added -mdnsinterface command line parameter to limit the MdnsBroadcaster to a single interface (e.g. matterbridge -bridge -mdnsinterface eth0). Matterbridge will validate the given interface and log a message if the interface is not available and will use all available interfaces.
 
 ### Changed
@@ -155,7 +159,7 @@ If you like this project and find it useful, please consider giving it a star on
 - [package]: Updated to use prettier and jest with the flat config
 - [matterbridge]: Updated dependencies
 - [matterbridgeDevice]: Updated ColorControl commandHandlers for improved compatibility with (https://github.com/Luligu/matterbridge-shelly)
-- [matterbridge]:  The "plugin disable" and "plugin remove" methods now also remove the registered devices from the bridge
+- [matterbridge]: The "plugin disable" and "plugin remove" methods now also remove the registered devices from the bridge
 - [matterbridge]: The "plugin add" method now loads, starts, and configures the plugin
 
 <a href="https://www.buymeacoffee.com/luligugithub">