+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 If you like this project and find it useful, please consider giving it a star on GitHub at https://github.com/Luligu/matterbridge and sponsoring it.
 
-## [1.3.14] - 2024-07-12
+## [1.3.15] - 2024-07-13
 
 ### Added
 
@@ -12,6 +12,7 @@ If you like this project and find it useful, please consider giving it a star on
 
 - [package]: Update dependencies.
 - [matterbridge]: Removed timeout on cleanup (saved 5 seconds on restart and shutdown).
+- [matterbridge]: Removed write cache and expired interval for node storage.
 - [frontend]: Frontend reduced timeout to reload page to 5 seconds.
 
 ### Fixed